[module4-task1] move modules
