+++ src/shared/modules/offer/index.ts
