+++ src/shared/modules/offer/offer-service.interface.ts
