+++ src/shared/modules/user/dto/create-user.dto.ts
@@ -1,4 +1,4 @@
-import { UserType } from "../../../../types/index.js";
+import { UserType } from "../../../types/index.js";
 
 export class CreateUserDto {
   public name: string;
