+++ src/shared/modules/user/user.container.ts
@@ -1,6 +1,6 @@
 import { Container } from "inversify";
 import { types } from "@typegoose/typegoose";
-import { Component } from "../../../types/index.js";
+import { Component } from "../../types/index.js";
 import { UserService } from "./user-service.interface.js";
 import { BaseUserService } from "./user.service.js";
 import { UserEntity, UserModel } from "./user.entity.js";
