+++ src/shared/modules/user/user.entity.ts
@@ -1,6 +1,6 @@
 import { defaultClasses, getModelForClass, prop, modelOptions } from "@typegoose/typegoose";
-import { User, UserType } from "../../../types/index.js";
-import { createSHA256 } from "../../../helpers/index.js";
+import { User, UserType } from "../../types/index.js";
+import { createSHA256 } from "../../helpers/index.js";
 
 // eslint-disable-next-line @typescript-eslint/no-unsafe-declaration-merging
 export interface UserEntity extends defaultClasses.Base {}
